+++ tests/Main.hs
@@ -26,40 +26,45 @@ main :: IO ()
 main = hspec $ do
   let mkSect level heading content ns =
         Tree.Node
-          (Section level ((), heading) ((), content))
+          (Section level heading () content ())
           ns
   describe "converting:" $ do
     it "empty document" $ do
       let src = ""
-          preface = ((), mempty)
+          preface = mempty
+          prefaceAnn = ()
           sections = []
       nodesToDocument (commonmarkToNodesWithSource [] src)
         `shouldBe` Document{..}
     it "spaces" $ do
       let src = "  \n\n  \n"
-          preface = ((), WithSource "  \n\n  \n" [])
+          preface = WithSource "  \n\n  \n" []
+          prefaceAnn = ()
           sections = []
       nodesToDocument (commonmarkToNodesWithSource [] src)
         `shouldBe` Document{..}
     it "paragraph" $ do
       let src = "x"
-          preface = ((), WithSource "x" [
-            Node (Just (PosInfo 1 1 1 1)) PARAGRAPH [text "x"] ])
+          preface = WithSource "x" [
+            Node (Just (PosInfo 1 1 1 1)) PARAGRAPH [text "x"] ]
+          prefaceAnn = ()
           sections = []
       nodesToDocument (commonmarkToNodesWithSource [] src)
         `shouldBe` Document{..}
     it "3 paragraphs" $ do
       let src = T.unlines ["","x","","","y","","z",""]
-          preface = ((), WithSource "\nx\n\n\ny\n\nz\n\n" [
+          preface = WithSource "\nx\n\n\ny\n\nz\n\n" [
             Node (Just (PosInfo 2 1 2 1)) PARAGRAPH [text "x"],
             Node (Just (PosInfo 5 1 5 1)) PARAGRAPH [text "y"],
-            Node (Just (PosInfo 7 1 7 1)) PARAGRAPH [text "z"] ])
+            Node (Just (PosInfo 7 1 7 1)) PARAGRAPH [text "z"] ]
+          prefaceAnn = ()
           sections = []
       nodesToDocument (commonmarkToNodesWithSource [] src)
         `shouldBe` Document{..}
     it "headers" $ do
       let src = T.unlines ["# 1", "", "## 2", "", "## 3"]
-          preface = ((), mempty)
+          preface = mempty
+          prefaceAnn = ()
           sections = [
             mkSect 1 (WithSource "# 1\n\n" [text "1"]) mempty [
               mkSect 2 (WithSource "## 2\n\n" [text "2"]) mempty [],
@@ -68,7 +73,8 @@ main = hspec $ do
         `shouldBe` Document{..}
     it "headers+content" $ do
       let src = T.unlines ["# 1", "", "## 2", "test", "## 3"]
-          preface = ((), mempty)
+          preface = mempty
+          prefaceAnn = ()
           sections = [
             mkSect 1 (WithSource "# 1\n\n" [text "1"]) mempty [
               mkSect 2 (WithSource "## 2\n" [text "2"])
@@ -79,7 +85,8 @@ main = hspec $ do
         `shouldBe` Document{..}
     it "preface+headers" $ do
       let src = T.unlines ["blah", "# 1", "", "## 2", "", "## 3"]
-          preface = ((), commonmarkToNodesWithSource [] "blah\n")
+          preface = commonmarkToNodesWithSource [] "blah\n"
+          prefaceAnn = ()
           sections = [
             mkSect 1 (WithSource "# 1\n\n" [text "1"]) mempty [
               mkSect 2 (WithSource "## 2\n\n" [text "2"]) mempty [],